Added psalm

+++ src/Executor/ParallelExecutor.php
@@ -40,7 +40,7 @@ final class ParallelExecutor extends AbstractExecutor
                 break;
             }
 
-            usleep(1000000 / $this->refreshRate);
+            usleep((int)(1000000 / $this->refreshRate));
         }
     }
 }
